+++ app/javascript/beyond_canvas/base.js
@@ -1,0 +1,3 @@
+import './initializers/buttons';
+import './initializers/flash';
+import './initializers/inputs';