+++ config/initializers/beyond_canvas/assets.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Rails.application.config.assets.precompile += %w( beyond_canvas/mailer.css )
-Rails.application.config.assets.precompile += %w( beyond_canvas.css )
-Rails.application.config.assets.precompile += %w( beyond_canvas.js )